pyproject.toml
[project]
name = "material-matcher"
version = "0.1.0"
description = "Excel材料价格匹配CLI工具"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "click>=8.0",
    "structlog>=23.0",
    "python-Levenshtein>=0.21",
    "pyyaml>=6.0",
    "requests>=2.28",
    "rich>=13.0",
]

[project.scripts]
material-matcher = "material_matcher.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
